restructures control panel and renames it to side panel
