+++ gui/src/components/readme.md
@@ -1,0 +1,6 @@
+# Components
+
+This module contains reusable components, e.g. an input component
+consisting of a single text edit and a button to apply the input.
+
+This module is visible to the whole crate.
