--- gui/src/controls/mod.rs
@@ -1,5 +1,0 @@
-mod tweak_control;
-mod next_round_control;
-mod game_status_display;
-mod resource_control;
-mod components;
