+++ gui/src/side_panel/game_status_display.rs
