+++ gui/src/side_panel/next_round_control.rs
