+++ gui/src/side_panel/readme.md
@@ -1,0 +1,6 @@
+# Side Panel
+
+This module contains the building blocks for the side panel
+
+The side panel itself is constructed in the `side_panel.rs` inside
+the parent module.
