+++ gui/src/side_panel/resource_control.rs
